refactor: Service 레이어 인터페이스화
apiserver 모듈 내 service 레이어 인터페이스화
