--- apiserver/controller/src/main/java/com/zipline/controller/user/TokenResponseDTO.java
@@ -1,4 +1,0 @@
-package com.zipline.controller.user;
-
-public class TokenResponseDTO {
-}
